LayoutTests/imported/w3c/web-platform-tests/css/css-text/animations/text-indent-interpolation-expected.txt


PASS CSS Transitions: property <text-indent> from neutral to [40px] at (-0.3) should be [1px] 
PASS CSS Transitions: property <text-indent> from neutral to [40px] at (0) should be [10px] 
PASS CSS Transitions: property <text-indent> from neutral to [40px] at (0.3) should be [19px] 
PASS CSS Transitions: property <text-indent> from neutral to [40px] at (0.6) should be [28px] 
PASS CSS Transitions: property <text-indent> from neutral to [40px] at (1) should be [40px] 
PASS CSS Transitions: property <text-indent> from neutral to [40px] at (1.5) should be [55px] 
PASS CSS Transitions with transition: all: property <text-indent> from neutral to [40px] at (-0.3) should be [1px] 
PASS CSS Transitions with transition: all: property <text-indent> from neutral to [40px] at (0) should be [10px] 
PASS CSS Transitions with transition: all: property <text-indent> from neutral to [40px] at (0.3) should be [19px] 
PASS CSS Transitions with transition: all: property <text-indent> from neutral to [40px] at (0.6) should be [28px] 
PASS CSS Transitions with transition: all: property <text-indent> from neutral to [40px] at (1) should be [40px] 
PASS CSS Transitions with transition: all: property <text-indent> from neutral to [40px] at (1.5) should be [55px] 
PASS CSS Animations: property <text-indent> from neutral to [40px] at (-0.3) should be [1px] 
PASS CSS Animations: property <text-indent> from neutral to [40px] at (0) should be [10px] 
PASS CSS Animations: property <text-indent> from neutral to [40px] at (0.3) should be [19px] 
PASS CSS Animations: property <text-indent> from neutral to [40px] at (0.6) should be [28px] 
PASS CSS Animations: property <text-indent> from neutral to [40px] at (1) should be [40px] 
PASS CSS Animations: property <text-indent> from neutral to [40px] at (1.5) should be [55px] 
PASS Web Animations: property <text-indent> from neutral to [40px] at (-0.3) should be [1px] 
FAIL Web Animations: property <text-indent> from neutral to [40px] at (0) should be [10px] assert_equals: expected "10px " but got "1px "
FAIL Web Animations: property <text-indent> from neutral to [40px] at (0.3) should be [19px] assert_equals: expected "19px " but got "12.7px "
FAIL Web Animations: property <text-indent> from neutral to [40px] at (0.6) should be [28px] assert_equals: expected "28px " but got "29.08px "
PASS Web Animations: property <text-indent> from neutral to [40px] at (1) should be [40px] 
FAIL Web Animations: property <text-indent> from neutral to [40px] at (1.5) should be [55px] assert_equals: expected "55px " but got "40px "
PASS CSS Transitions: property <text-indent> from [initial] to [20px] at (-0.3) should be [-6px] 
PASS CSS Transitions: property <text-indent> from [initial] to [20px] at (0) should be [0px] 
PASS CSS Transitions: property <text-indent> from [initial] to [20px] at (0.3) should be [6px] 
PASS CSS Transitions: property <text-indent> from [initial] to [20px] at (0.6) should be [12px] 
PASS CSS Transitions: property <text-indent> from [initial] to [20px] at (1) should be [20px] 
PASS CSS Transitions: property <text-indent> from [initial] to [20px] at (1.5) should be [30px] 
PASS CSS Transitions with transition: all: property <text-indent> from [initial] to [20px] at (-0.3) should be [-6px] 
PASS CSS Transitions with transition: all: property <text-indent> from [initial] to [20px] at (0) should be [0px] 
PASS CSS Transitions with transition: all: property <text-indent> from [initial] to [20px] at (0.3) should be [6px] 
PASS CSS Transitions with transition: all: property <text-indent> from [initial] to [20px] at (0.6) should be [12px] 
PASS CSS Transitions with transition: all: property <text-indent> from [initial] to [20px] at (1) should be [20px] 
PASS CSS Transitions with transition: all: property <text-indent> from [initial] to [20px] at (1.5) should be [30px] 
PASS CSS Animations: property <text-indent> from [initial] to [20px] at (-0.3) should be [-6px] 
PASS CSS Animations: property <text-indent> from [initial] to [20px] at (0) should be [0px] 
PASS CSS Animations: property <text-indent> from [initial] to [20px] at (0.3) should be [6px] 
PASS CSS Animations: property <text-indent> from [initial] to [20px] at (0.6) should be [12px] 
PASS CSS Animations: property <text-indent> from [initial] to [20px] at (1) should be [20px] 
PASS CSS Animations: property <text-indent> from [initial] to [20px] at (1.5) should be [30px] 
PASS Web Animations: property <text-indent> from [initial] to [20px] at (-0.3) should be [-6px] 
PASS Web Animations: property <text-indent> from [initial] to [20px] at (0) should be [0px] 
PASS Web Animations: property <text-indent> from [initial] to [20px] at (0.3) should be [6px] 
PASS Web Animations: property <text-indent> from [initial] to [20px] at (0.6) should be [12px] 
PASS Web Animations: property <text-indent> from [initial] to [20px] at (1) should be [20px] 
PASS Web Animations: property <text-indent> from [initial] to [20px] at (1.5) should be [30px] 
PASS CSS Transitions: property <text-indent> from [inherit] to [20px] at (-0.3) should be [85px] 
PASS CSS Transitions: property <text-indent> from [inherit] to [20px] at (0) should be [70px] 
PASS CSS Transitions: property <text-indent> from [inherit] to [20px] at (0.3) should be [55px] 
PASS CSS Transitions: property <text-indent> from [inherit] to [20px] at (0.6) should be [40px] 
PASS CSS Transitions: property <text-indent> from [inherit] to [20px] at (1) should be [20px] 
PASS CSS Transitions: property <text-indent> from [inherit] to [20px] at (1.5) should be [-5px] 
PASS CSS Transitions with transition: all: property <text-indent> from [inherit] to [20px] at (-0.3) should be [85px] 
PASS CSS Transitions with transition: all: property <text-indent> from [inherit] to [20px] at (0) should be [70px] 
PASS CSS Transitions with transition: all: property <text-indent> from [inherit] to [20px] at (0.3) should be [55px] 
PASS CSS Transitions with transition: all: property <text-indent> from [inherit] to [20px] at (0.6) should be [40px] 
PASS CSS Transitions with transition: all: property <text-indent> from [inherit] to [20px] at (1) should be [20px] 
PASS CSS Transitions with transition: all: property <text-indent> from [inherit] to [20px] at (1.5) should be [-5px] 
PASS CSS Animations: property <text-indent> from [inherit] to [20px] at (-0.3) should be [85px] 
PASS CSS Animations: property <text-indent> from [inherit] to [20px] at (0) should be [70px] 
PASS CSS Animations: property <text-indent> from [inherit] to [20px] at (0.3) should be [55px] 
PASS CSS Animations: property <text-indent> from [inherit] to [20px] at (0.6) should be [40px] 
PASS CSS Animations: property <text-indent> from [inherit] to [20px] at (1) should be [20px] 
PASS CSS Animations: property <text-indent> from [inherit] to [20px] at (1.5) should be [-5px] 
FAIL Web Animations: property <text-indent> from [inherit] to [20px] at (-0.3) should be [85px] assert_equals: expected "85px " but got "7px "
FAIL Web Animations: property <text-indent> from [inherit] to [20px] at (0) should be [70px] assert_equals: expected "70px " but got "7px "
FAIL Web Animations: property <text-indent> from [inherit] to [20px] at (0.3) should be [55px] assert_equals: expected "55px " but got "10.9px "
FAIL Web Animations: property <text-indent> from [inherit] to [20px] at (0.6) should be [40px] assert_equals: expected "40px " but got "16.36px "
PASS Web Animations: property <text-indent> from [inherit] to [20px] at (1) should be [20px] 
FAIL Web Animations: property <text-indent> from [inherit] to [20px] at (1.5) should be [-5px] assert_equals: expected "- 5px " but got "20px "
PASS CSS Transitions: property <text-indent> from [unset] to [20px] at (-0.3) should be [85px] 
PASS CSS Transitions: property <text-indent> from [unset] to [20px] at (0) should be [70px] 
PASS CSS Transitions: property <text-indent> from [unset] to [20px] at (0.3) should be [55px] 
PASS CSS Transitions: property <text-indent> from [unset] to [20px] at (0.6) should be [40px] 
PASS CSS Transitions: property <text-indent> from [unset] to [20px] at (1) should be [20px] 
PASS CSS Transitions: property <text-indent> from [unset] to [20px] at (1.5) should be [-5px] 
PASS CSS Transitions with transition: all: property <text-indent> from [unset] to [20px] at (-0.3) should be [85px] 
PASS CSS Transitions with transition: all: property <text-indent> from [unset] to [20px] at (0) should be [70px] 
PASS CSS Transitions with transition: all: property <text-indent> from [unset] to [20px] at (0.3) should be [55px] 
PASS CSS Transitions with transition: all: property <text-indent> from [unset] to [20px] at (0.6) should be [40px] 
PASS CSS Transitions with transition: all: property <text-indent> from [unset] to [20px] at (1) should be [20px] 
PASS CSS Transitions with transition: all: property <text-indent> from [unset] to [20px] at (1.5) should be [-5px] 
PASS CSS Animations: property <text-indent> from [unset] to [20px] at (-0.3) should be [85px] 
PASS CSS Animations: property <text-indent> from [unset] to [20px] at (0) should be [70px] 
PASS CSS Animations: property <text-indent> from [unset] to [20px] at (0.3) should be [55px] 
PASS CSS Animations: property <text-indent> from [unset] to [20px] at (0.6) should be [40px] 
PASS CSS Animations: property <text-indent> from [unset] to [20px] at (1) should be [20px] 
PASS CSS Animations: property <text-indent> from [unset] to [20px] at (1.5) should be [-5px] 
FAIL Web Animations: property <text-indent> from [unset] to [20px] at (-0.3) should be [85px] assert_equals: expected "85px " but got "7px "
FAIL Web Animations: property <text-indent> from [unset] to [20px] at (0) should be [70px] assert_equals: expected "70px " but got "7px "
FAIL Web Animations: property <text-indent> from [unset] to [20px] at (0.3) should be [55px] assert_equals: expected "55px " but got "10.9px "
FAIL Web Animations: property <text-indent> from [unset] to [20px] at (0.6) should be [40px] assert_equals: expected "40px " but got "16.36px "
PASS Web Animations: property <text-indent> from [unset] to [20px] at (1) should be [20px] 
FAIL Web Animations: property <text-indent> from [unset] to [20px] at (1.5) should be [-5px] assert_equals: expected "- 5px " but got "20px "
PASS CSS Transitions: property <text-indent> from [0px] to [50px] at (-0.3) should be [-15px] 
PASS CSS Transitions: property <text-indent> from [0px] to [50px] at (0) should be [0] 
PASS CSS Transitions: property <text-indent> from [0px] to [50px] at (0.3) should be [15px] 
PASS CSS Transitions: property <text-indent> from [0px] to [50px] at (0.6) should be [30px] 
PASS CSS Transitions: property <text-indent> from [0px] to [50px] at (1) should be [50px] 
PASS CSS Transitions: property <text-indent> from [0px] to [50px] at (1.5) should be [75px] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px] to [50px] at (-0.3) should be [-15px] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px] to [50px] at (0) should be [0] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px] to [50px] at (0.3) should be [15px] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px] to [50px] at (0.6) should be [30px] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px] to [50px] at (1) should be [50px] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px] to [50px] at (1.5) should be [75px] 
PASS CSS Animations: property <text-indent> from [0px] to [50px] at (-0.3) should be [-15px] 
PASS CSS Animations: property <text-indent> from [0px] to [50px] at (0) should be [0] 
PASS CSS Animations: property <text-indent> from [0px] to [50px] at (0.3) should be [15px] 
PASS CSS Animations: property <text-indent> from [0px] to [50px] at (0.6) should be [30px] 
PASS CSS Animations: property <text-indent> from [0px] to [50px] at (1) should be [50px] 
PASS CSS Animations: property <text-indent> from [0px] to [50px] at (1.5) should be [75px] 
PASS Web Animations: property <text-indent> from [0px] to [50px] at (-0.3) should be [-15px] 
PASS Web Animations: property <text-indent> from [0px] to [50px] at (0) should be [0] 
PASS Web Animations: property <text-indent> from [0px] to [50px] at (0.3) should be [15px] 
PASS Web Animations: property <text-indent> from [0px] to [50px] at (0.6) should be [30px] 
PASS Web Animations: property <text-indent> from [0px] to [50px] at (1) should be [50px] 
PASS Web Animations: property <text-indent> from [0px] to [50px] at (1.5) should be [75px] 
PASS CSS Transitions: property <text-indent> from [0px hanging] to [50px hanging] at (-0.3) should be [-15px hanging] 
PASS CSS Transitions: property <text-indent> from [0px hanging] to [50px hanging] at (0) should be [0 hanging] 
PASS CSS Transitions: property <text-indent> from [0px hanging] to [50px hanging] at (0.3) should be [15px hanging] 
PASS CSS Transitions: property <text-indent> from [0px hanging] to [50px hanging] at (0.6) should be [30px hanging] 
PASS CSS Transitions: property <text-indent> from [0px hanging] to [50px hanging] at (1) should be [50px hanging] 
PASS CSS Transitions: property <text-indent> from [0px hanging] to [50px hanging] at (1.5) should be [75px hanging] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px hanging] to [50px hanging] at (-0.3) should be [-15px hanging] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px hanging] to [50px hanging] at (0) should be [0 hanging] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px hanging] to [50px hanging] at (0.3) should be [15px hanging] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px hanging] to [50px hanging] at (0.6) should be [30px hanging] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px hanging] to [50px hanging] at (1) should be [50px hanging] 
PASS CSS Transitions with transition: all: property <text-indent> from [0px hanging] to [50px hanging] at (1.5) should be [75px hanging] 
PASS CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (-0.3) should be [-15px hanging] 
PASS CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0) should be [0 hanging] 
PASS CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.3) should be [15px hanging] 
PASS CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.6) should be [30px hanging] 
PASS CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1) should be [50px hanging] 
PASS CSS Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1.5) should be [75px hanging] 
PASS Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (-0.3) should be [-15px hanging] 
PASS Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0) should be [0 hanging] 
PASS Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.3) should be [15px hanging] 
PASS Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (0.6) should be [30px hanging] 
PASS Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1) should be [50px hanging] 
PASS Web Animations: property <text-indent> from [0px hanging] to [50px hanging] at (1.5) should be [75px hanging] 
FAIL CSS Transitions: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (-0.3) should be [-15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0) should be [0 hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.3) should be [15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.6) should be [30px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1) should be [50px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1.5) should be [75px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (-0.3) should be [-15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0) should be [0 hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.3) should be [15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.6) should be [30px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1) should be [50px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1.5) should be [75px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (-0.3) should be [-15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0) should be [0 hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.3) should be [15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.6) should be [30px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1) should be [50px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1.5) should be [75px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (-0.3) should be [-15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0) should be [0 hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.3) should be [15px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (0.6) should be [30px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1) should be [50px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px hanging each-line] to [50px each-line hanging] at (1.5) should be [75px hanging each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (1) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (1) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [0px each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [0px each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [0px each-line] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (-0.3) should be [0px each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0) should be [0px each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.3) should be [0px each-line] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (0.6) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px each-line] to [50px hanging] at (1.5) should be [50px hanging] assert_true: 'from' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (1) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (1) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Transitions with transition: all: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [0px] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [0px] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [0px] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL CSS Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (-0.3) should be [0px] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0) should be [0px] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.3) should be [0px] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (0.6) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
FAIL Web Animations: property <text-indent> from [0px] to [50px each-line hanging] at (1.5) should be [50px each-line hanging] assert_true: 'to' value should be supported expected true got false
